// src/app/core/models/clinical-notes/recall.model.ts
export interface CreateRecallDto {
  hygieneProvider: string;
  examProvider: string;
  medication: string;
  medicalHx?: string;
  IntraOrExtraoral?: string;
  parq?: boolean;
  healthHxCheck?: boolean;
  bp?: string;
  pulse?: string;
  examType?: string;
  cleaningType?: string;
  periapicalXray?: string;
  bitewingXray?: string;
  otherXray?: string;
  fluroide?: string;
  periodontalProbe?: string;
  otherTreatment?: string;
  oralHygieneInstruction?: string;
  treatmentNotes?: string;
  oralCancerScreen?: string;
  oralHygiene?: string;
  stainH?: boolean;
  stainM?: boolean;
  stainL?: boolean;
  subGlacH?: boolean;
  subGlacM?: boolean;
  subGlacL?: boolean;
  supraClacH?: boolean;
  supraClacM?: boolean;
  supraClacL?: boolean;
  bleedingH?: boolean;
  bleedingM?: boolean;
  bleedingL?: boolean;
  prioCaseI?: boolean;
  prioCaseII?: boolean;
  prioCaseIII?: boolean;
  prioCaseIV?: boolean;
  perioScreeningTxt1?: string;
  perioScreeningTxt2?: string;
  perioScreeningTxt3?: string;
  perioScreeningTxt4?: string;
  perioScreeningTxt5?: string;
  perioScreeningTxt6?: string;
  recallInterval?: string;
  due?: string;
  time?: string;
  cheifComplaint?: string;
  lips?: string;
  buccal?: string;
  vestibules?: string;
  tongue?: string;
  floorOfMouth?: string;
  palate?: string;
  gingiva?: string;
  torus?: string;
  tmjPain?: string;
  crepitus?: string;
  snap?: string;
  deviate?: string;
  maxOpening?: string;
}